Fix incorrect format strings and format arguments.
Even though int and long happen to be of the same size on the Amiga we
should still use valid C when possible.

+++ Tools/CryptPasswords/CryptPasswords.c
@@ -36,7 +36,8 @@ void convertUser(int userId) {
     return;
   }
   if(statbuf.st_size != sizeof(struct User)) {
-    printf("Size of user's data file is %d, was expecting %d.\n", statbuf.st_size, sizeof(struct User));
+    printf("Size of user's data file is %d, was expecting %lu.\n",
+           statbuf.st_size, (unsigned long)sizeof(struct User));
     printf("Are you running the right version of CryptPasswords?\n");
     return;
   }
